requirements.txt
Package         Version
--------------- -------
async-timeout   4.0.2
confluent-kafka 2.1.1
dnspython       2.3.0
install         1.3.5
kafka-python    2.0.2
pip             22.0.2
pymongo         4.3.3
redis           4.5.5
setuptools      59.6.0
